drfecommerce/tests/factories/brand.py
import factory

from drfecommerce.product.models import Brand


class BrandFactory(factory.django.DjangoModelFactory):
    class Meta:
        model = Brand

    name = factory.Sequence(lambda n: "Brand_%d" % n)
