pyproject.toml
[tool.poetry]
name = "django-drf-ecommerce"
version = "0.1.0"
description = "Learning Django REST Framework for E-commerce"
authors = ["Socheath Horl <horl.socheath@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
pytest = "^7.4.3"
pytest-django = "^4.7.0"
black = "^23.11.0"
flake8 = "^6.1.0"
django-mptt = "^0.15.0"
drf-spectacular = "^0.26.5"
coverage = "^7.3.2"
pytest-cov = "^4.1.0"
pytest-factoryboy = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
